requirements.txt
altair==6.0.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
filetype==1.2.0
fonttools==4.61.1
fqdn==1.5.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
lap==0.5.12
lark==1.3.1
MarkupSafe==2.1.5
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.1
notebook_shim==0.2.4
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
platformdirs==4.5.1
polars==1.36.1
polars-runtime-32==1.36.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.2.1
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
roboflow==1.2.11
rpds-py==0.30.0
scipy==1.16.3
Send2Trash==2.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
stack-data==0.6.3
streamlit==1.52.2
sympy==1.13.1
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.248
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.6.2
watchdog==6.0.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
